# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Imari API
  version: 0.1.0
  description: Personalised financial literacy platform — ALU pilot.
servers:
  - url: https://api.imari.app/v1
security:
  - bearer: []
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
...
    Credential:
      type: object
      properties:
        id: { type: string, pattern: '^IMR-[A-Z0-9]{4}-[A-Z0-9]{4}$' }
        competency: { type: string }
        holderName: { type: string }
        hash: { type: string }
        txHash: { type: [string, 'null'] }
        block: { type: [integer, 'null'] }
        status: { type: string, enum: [pending, valid, revoked] }
        issuedAt: { type: string, format: date-time }
    Verification:
      type: object
      properties:
        outcome: { type: string, enum: [valid, revoked, mismatch, not_found] }
        credential: { $ref: '#/components/schemas/Credential' }
paths:
...
  /modules/{slug}:
    get:
      parameters: [{ name: slug, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: Module with lessons, activity and quiz (answers stripped) } }
  /modules/{slug}/quiz:
    post:
      parameters: [{ name: slug, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: '{ score, passed }' } }
  /modules/{slug}/credential:
    post:
      parameters: [{ name: slug, in: path, required: true, schema: { type: string } }]
      responses:
        '201': { description: Issued (status may be pending until mined), content: { application/json: { schema: { $ref: '#/components/schemas/Credential' } } } }
        '409': { description: COMPETENCY_NOT_MET }
  /verify/{credentialId}:
    get:
      security: []
      parameters:
        - { name: credentialId, in: path, required: true, schema: { type: string } }
        - { name: hash, in: query, schema: { type: string } }
      responses: { '200': { description: Result, content: { application/json: { schema: { $ref: '#/components/schemas/Verification' } } } } }
...